REFACTOR: add custom hook for the TalkToUsForm
