+++ src/components/TalkToUs/TalkToUsForm.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { useTalkToUsForm } from "@/hook/useTalkToUsForm";
 import { schemaTalkToUs, SchemaTalkToUsProps } from "@/model/schemas/schemaTalkToUs";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { resolve } from "path"; // remover importações não usadas
@@ -7,45 +8,20 @@ import { useState } from "react";
 import { useForm } from "react-hook-form";
 
 const TalkToUsForm = () => {
+    const {
+       register, 
+       onSubmit, 
+       handleSubmit,
+       formState: {
+        errors,
+        isSubmitted,
+        isSubmitting,
+        isSubmitSuccessful
+       },
+       showFormMessages,
+       formErrorMessage
+    } = useTalkToUsForm()
 
-    //sugestão de melhoria : Modularize a lógica de envio em um hook para melhorar a manutenibilidade, testabilidade e clareza do código, tornando o componente TalkToUsForm mais simples e focado apenas na renderização.
-
-    const [formErrorMessage, setFormErrorMessage] = useState<string | null>(null)
-    const [showFormMessages, setShowFormMessages] = useState<boolean>(false)
-
-    const { 
-        register, 
-        handleSubmit, 
-        reset, 
-        formState: {
-            errors, 
-            isSubmitting, 
-            isSubmitSuccessful, 
-            isSubmitted
-        } 
-    } = useForm<SchemaTalkToUsProps>({
-        resolver: yupResolver(schemaTalkToUs)
-    })
-    const onSubmit = async (data: SchemaTalkToUsProps) => {
-        //TODO enviar dados para a API
-
-        setFormErrorMessage(null)
-        setShowFormMessages(false)
-
-        try {
-            await new Promise(resolve => setTimeout(resolve, 2000))
-            console.log(data)
-
-            reset()
-            setShowFormMessages(true)
-            setTimeout(() => setShowFormMessages(false), 3000)
-        } catch(error) {
-            setFormErrorMessage("Ocorreu um erro ao enviar o formulário")
-            setShowFormMessages(true)
-            setTimeout(() => setShowFormMessages(false), 3000)
-        }
-
-    }
     return ( 
         <form onSubmit={handleSubmit(onSubmit)} className="mx-auto w-full max-w-[551px] space-y-4 lg:space-y-10 mt-8 lg:mt-0"> 
             <fieldset>
